data_consumer/requirements.txt
kafka-python-ng
pyspark
pymongo
pandas
psycopg2-binary